+++ src/components/Channel/Channel.tsx
@@ -377,7 +377,7 @@ const ChannelInner = <
 
   const markReadThrottled = throttle(markRead, 500, { leading: true, trailing: true });
 
-  const handleEvent = (event: Event<StreamChatGenerics>) => {
+  const handleEvent = async (event: Event<StreamChatGenerics>) => {
     if (event.message) {
       dispatch({
         channel,
@@ -432,6 +432,18 @@ const ChannelInner = <
           latestMessageDatesByChannels[cid] = messageDate;
         }
       }
+    }
+
+    if (event.type === 'user.deleted') {
+      const oldestID = channel.state?.messages?.[0]?.id;
+
+      /**
+       * As the channel state is not normalized we re-fetch the channel data. Thus, we avoid having to search for user references in the channel state.
+       */
+      await channel.query({
+        messages: { id_lt: oldestID, limit: DEFAULT_NEXT_CHANNEL_PAGE_SIZE },
+        watchers: { limit: DEFAULT_NEXT_CHANNEL_PAGE_SIZE },
+      });
     }
 
     throttledCopyStateFromChannel();
@@ -670,8 +682,26 @@ const ChannelInner = <
         messageResponse = await channel.sendMessage(messageData);
       }
 
-      // replace it after send is completed
-      if (messageResponse?.message) {
+      let existingMessage;
+      for (let i = channel.state.messages.length - 1; i >= 0; i--) {
+        const msg = channel.state.messages[i];
+        if (msg.id === messageData.id) {
+          existingMessage = msg;
+          break;
+        }
+      }
+
+      const responseTimestamp = new Date(messageResponse?.message?.updated_at || 0).getTime();
+      const existingMessageTimestamp = existingMessage?.updated_at?.getTime() || 0;
+      const responseIsTheNewest = responseTimestamp > existingMessageTimestamp;
+
+      // Replace the message payload after send is completed
+      // We need to check for the newest message payload, because on slow network, the response can arrive later than WS events message.new, message.updated.
+      // Always override existing message in status "sending"
+      if (
+        messageResponse?.message &&
+        (responseIsTheNewest || existingMessage?.status === 'sending')
+      ) {
         updateMessage({
           ...messageResponse.message,
           status: 'received',